requirements.txt
numpy==1.19.2
opencv-contrib-python==4.4.0.42
Pillow==7.2.0
